requirements.txt
# Core dependencies
streamlit>=1.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0

# ML/AI dependencies
torch>=2.0.0
transformers>=4.34.0
accelerate>=0.24.0
google-generativeai>=0.3.0

# Database and API
supabase>=2.0.0
psycopg2-binary>=2.9.0

# Async support
async-timeout>=4.0.0
asyncio>=3.4.3

# Utility
python-dotenv>=1.0.0
python-dateutil>=2.8.2
requests>=2.31.0
tqdm>=4.66.0
